+++ diskette/views/__init__.py
@@ -1,6 +1,7 @@
-from .dump import DumpFileAdminDownloadView
+from .dump import DumpFileAdminDownloadView, DumpLogAdminView
 
 
 __all__ = [
     "DumpFileAdminDownloadView",
+    "DumpLogAdminView",
 ]